Implementando verificação ao deletar endereço

+++ src/main/java/br/com/dbccompany/time7/gestaodeensino/service/EnderecoService.java
@@ -5,7 +5,9 @@ import br.com.dbccompany.time7.gestaodeensino.dto.EnderecoDTO;
 import br.com.dbccompany.time7.gestaodeensino.dto.EnderecoUpdateDTO;
 import br.com.dbccompany.time7.gestaodeensino.entity.EnderecoEntity;
 import br.com.dbccompany.time7.gestaodeensino.exceptions.RegraDeNegocioException;
+import br.com.dbccompany.time7.gestaodeensino.repository.AlunoRepository;
 import br.com.dbccompany.time7.gestaodeensino.repository.EnderecoRepository;
+import br.com.dbccompany.time7.gestaodeensino.repository.ProfessorRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,6 +18,10 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class EnderecoService {
     private final EnderecoRepository enderecoRepository;
+
+    private final AlunoRepository alunoRepository;
+
+    private final ProfessorRepository professorRepository;
     private final ObjectMapper objectMapper;
 
     public EnderecoDTO findById(Integer idEndereco) throws RegraDeNegocioException {
@@ -52,17 +58,24 @@ public class EnderecoService {
     }
 
     public void delete(Integer idEndereco) throws RegraDeNegocioException {
-        log.info("Removendo endereço");
+        Integer count = professorRepository.countProfessorEntityByIdEndereco(idEndereco);
+        count += alunoRepository.countProfessorEntityByIdEndereco(idEndereco);
 
-        EnderecoEntity enderecoDelete = enderecoRepository.findById(idEndereco).
-                orElseThrow(() -> new RegraDeNegocioException("Endereço não encontrado"));
+        if (count == 0) {
+            log.info("Removendo endereço");
 
-        enderecoDelete.setProfessorEntity(enderecoDelete.getProfessorEntity());
-        enderecoDelete.setAlunoEntity(enderecoDelete.getAlunoEntity());
+            EnderecoEntity enderecoDelete = enderecoRepository.findById(idEndereco).
+                    orElseThrow(() -> new RegraDeNegocioException("Endereço não encontrado"));
 
-        enderecoRepository.delete(enderecoDelete);
+            enderecoDelete.setProfessorEntity(enderecoDelete.getProfessorEntity());
+            enderecoDelete.setAlunoEntity(enderecoDelete.getAlunoEntity());
 
-        log.info("Endereço removido");
+            enderecoRepository.delete(enderecoDelete);
+
+            log.info("Endereço removido");
+        } else {
+            throw new RegraDeNegocioException("Endereço está sendo utilizado");
+        }
     }
 
     public EnderecoEntity createToEntity(EnderecoCreateDTO enderecoCreateDTO) {
